Warning when get_batch can be very slow
When entering loop 97, if len(observations) is not much larger than batch, it is extremely slow (minutes), due to the way the algorithm works. This was very confusing to me so I thought it might be valuable to add a warning, although I'm not sure if it's needed and I'm not sure if this is the best way. Happy to go with what you think.

This can wait until the optimiser update, cheers.;