Independent save and restore of model components
This change introduces the mixin SavableComponent that model
components that are to be saved or restored independently from the
model should extend from.

The Model calls register_saver_ops of each savable component when
computation graph is constructed and save operations can be added to
it.

Note that with this change the save and restore of the savable
components need to be called explicitly. Model's save, restore and
checkpoint functionality does not delegate to the savable components'
save and restore methods.

A simple recipe to train/preload embeddings can be found in
TestModelSaveRestore.test_pretrain_network