Introduce save_component and restore_component
This change introduces the methods save_component and restore_component
as shorthands to save and restore of the actual objects.

The methods accept a `component_name` argument which should match the
value of one of the model's components. The session used to carry on
the operation is the model's session.